Add `NtWriteFile` to `foreign_items`
